bd_app/services/auth.py
import os
from datetime import datetime, timedelta
from typing import Optional

from jose import JWTError, jwt
from passlib.context import CryptContext
from pydantic import BaseModel

pwd_context = CryptContext(schemes=["bcrypt"], deprecated="auto")

SECRET_KEY = os.getenv("JWT_SECRET_KEY", "changeme")
ALGORITHM = os.getenv("JWT_ALGORITHM", "HS256")
ACCESS_TOKEN_EXPIRE_MINUTES = int(os.getenv("JWT_ACCESS_TOKEN_EXPIRE_MINUTES", "30"))

class User(BaseModel):
    email: str
    password_hash: str
    role: str = "user"

_fake_users: dict[str, User] = {}

def hash_password(password: str) -> str:
    return pwd_context.hash(password)

def verify_password(plain_password: str, hashed_password: str) -> bool:
    return pwd_context.verify(plain_password, hashed_password)

def create_access_token(data: dict, expires_delta: Optional[timedelta] = None) -> str:
    to_encode = data.copy()
    expire = datetime.utcnow() + (expires_delta or timedelta(minutes=ACCESS_TOKEN_EXPIRE_MINUTES))
    to_encode.update({"exp": expire})
    return jwt.encode(to_encode, SECRET_KEY, algorithm=ALGORITHM)

def verify_token(token: str) -> dict:
    try:
        return jwt.decode(token, SECRET_KEY, algorithms=[ALGORITHM])
    except JWTError as e:
        raise ValueError("Invalid token") from e

def create_user(user_data: dict) -> User:
    password_hash = hash_password(user_data["password"])
    user = User(email=user_data["email"], password_hash=password_hash)
    _fake_users[user.email] = user
    return user

def authenticate_user(email: str, password: str) -> Optional[User]:
    user = _fake_users.get(email)
    if not user:
        return None
    if not verify_password(password, user.password_hash):
        return None
    return user

def get_current_user(token: str) -> User:
    payload = verify_token(token)
    email = payload.get("sub")
    if email is None:
        raise ValueError("Invalid token payload")
    user = _fake_users.get(email)
    if user is None:
        raise ValueError("User not found")
    return user
